Move Audio Plugin classes to domain.plugins
