+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/domain/content/TakeActions.kt
@@ -3,7 +3,7 @@ package org.wycliffeassociates.otter.common.domain.content
 import io.reactivex.Single
 import org.wycliffeassociates.otter.common.audio.wav.EMPTY_WAVE_FILE_SIZE
 import org.wycliffeassociates.otter.common.audio.wav.IWaveFileCreator
-import org.wycliffeassociates.otter.common.data.PluginParameters
+import org.wycliffeassociates.otter.common.domain.plugins.PluginParameters
 import org.wycliffeassociates.otter.common.data.model.MimeType
 import org.wycliffeassociates.otter.common.data.workbook.AssociatedAudio
 import org.wycliffeassociates.otter.common.data.workbook.Take
