+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/domain/plugins/LaunchPlugin.kt
@@ -2,7 +2,6 @@ package org.wycliffeassociates.otter.common.domain.plugins
 
 import io.reactivex.Maybe
 import io.reactivex.Single
-import org.wycliffeassociates.otter.common.data.PluginParameters
 import org.wycliffeassociates.otter.common.persistence.repositories.IAudioPluginRepository
 import org.wycliffeassociates.otter.common.persistence.repositories.PluginType
 import java.io.File
