+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/audioplugin/parser/ParsedAudioPluginDataMapper.kt
@@ -1,6 +1,6 @@
 package org.wycliffeassociates.otter.jvm.workbookapp.audioplugin.parser
 
-import org.wycliffeassociates.otter.common.data.config.AudioPluginData
+import org.wycliffeassociates.otter.common.domain.plugins.AudioPluginData
 import java.io.File
 
 /** @param osName optionally overrides the value from System.getProperty("os.name") */
